+++ co-log-polysemy/CHANGELOG.md
@@ -3,10 +3,16 @@
 `co-log-polysemy` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
-## Unreleased
+## 0.0.1.0 — Jan 19, 2020
 
 * [#121](https://github.com/kowainik/co-log/issues/121):
   Add `LogAction` and interpreter that works directly with the `Sem` monad.
+  (by [@chshersh](https://github.com/chshersh))
+* [#148](https://github.com/kowainik/co-log/issues/148):
+  Support GHC-8.8.2.
+  (by [@chshersh](https://github.com/chshersh))
+* Bump up `polysemy` to the latest version.
+  (by [@Avi-D-coder](https://github.com/Avi-D-coder))
 
 ## 0.0.0.0 — May 5, 2019